+++ openeogeotrellis/user_defined_process_repository.py
@@ -20,7 +20,7 @@ class ZooKeeperUserDefinedProcessRepository(UserDefinedProcesses):
 
     _log = logging.getLogger(__name__)
 
-    def __init__(self, hosts: List[str], root: str = "/openeo/udps", zk_client_reuse: bool = False):
+    def __init__(self, hosts: List[str], root: str = "/openeo/udps", *, zk_client_reuse: bool = False):
         self._hosts = ",".join(hosts)
         self._root = root
         self._zk_client_reuse = zk_client_reuse
@@ -28,13 +28,11 @@ class ZooKeeperUserDefinedProcessRepository(UserDefinedProcesses):
 
     @staticmethod
     def _serialize(spec: dict) -> bytes:
-        return json.dumps({
-            'specification': spec
-        }).encode()
+        return json.dumps({"specification": spec}).encode("utf8")
 
     @staticmethod
     def _deserialize(data: bytes) -> dict:
-        return json.loads(data.decode())
+        return json.loads(data.decode("utf8"))
 
     def save(self, user_id: str, process_id: str, spec: dict) -> None:
         with self._zk_client() as zk:
@@ -141,7 +139,7 @@ class InMemoryUserDefinedProcessRepository(UserDefinedProcesses):
 
 
 def main():
-    repo = ZooKeeperUserDefinedProcessRepository(hosts=ConfigParams().zookeepernodes)
+    repo = ZooKeeperUserDefinedProcessRepository(hosts=ConfigParams().zookeepernodes, zk_client_reuse=True)
 
     user_id = 'vdboschj'
     process_graph_id = 'evi'